Make json desrialization faster, by using long types instead of int so we use the supported convertion between long to Nullable<long>, instead of adding support for TryGet method for converting long to Nullable<int>.

+++ src/Raven.Server/Documents/Versioning/VersioningConfigurationCollection.cs
@@ -2,7 +2,7 @@ namespace Raven.Server.Documents.Versioning
 {
     public class VersioningConfigurationCollection
     {
-        public int? MaxRevisions { get; set; }
+        public long? MaxRevisions { get; set; }
 
         public bool Active { get; set; }
